requirements.txt
pyyaml>=6.0
google-cloud-storage>=2.0.0
google-cloud-logging>=3.0.0
click>=8.0.0
pytest>=7.0.0
flake8>=6.0.0
black>=23.0.0
